seperate the post models and managers into their respective dirs
